make all the parameters short

+++ main.c
@@ -4,27 +4,27 @@
 
 
 // ===========================> Functions Prototype <===============================
-void conv1d(const int *data, const int *filter, int *map_out, const int *bias, int filter_size,
+void conv1d(const short *data, const short *filter, short *map_out, const short *bias, int filter_size,
             int input_len, int input_depth, int output_len, int n_filter, int strides, int relu, int padding);
 
-void batch_normalization(const int *data, const int *gamma, const int *beta, const int *mean, const int *var,
-                         int *map_out, int input_len, int input_depth);
+void batch_normalization(const short *data, const short *gamma, const short *beta, const short *mean, const short *var,
+                         short *map_out, int input_len, int input_depth);
 
-void skip_add(const int *data, int *map_out, int input_len);
+void skip_add(const short *data, short *map_out, int input_len);
 
-void forward_propagation(int *data);
+void forward_propagation(short *data);
 // =================================================================================
 
 int main() {
     // allocate memory in CPU for calculation
-    int *eeg_input;
+    short *eeg_input;
     eeg_input = input_array;
     forward_propagation(eeg_input);
     return 0;
 }
 
 
-void conv1d(const int *data, const int *filter, int *map_out,const int *bias, const int filter_size,
+void conv1d(const short *data, const short *filter, short *map_out,const short *bias, const int filter_size,
             const int input_len, const int input_depth, const int output_len, const int n_filter, const int strides,
             const int relu, const int padding) {
     int sum;
@@ -45,12 +45,12 @@ void conv1d(const int *data, const int *filter, int *map_out,const int *bias, co
             sum += bias[w_n];
             if (sum < 0 && relu)
                 sum = 0; // Relu
-            mem2d(map_out, output_len, w_n, start_index/strides) = sum;
+            mem2d(map_out, output_len, w_n, start_index/strides) = (short) sum;
         }
     }
 }
 
-void max1d(const int *data, int *map_out, const int input_len, const int input_depth,
+void max1d(const short *data, short *map_out, const int input_len, const int input_depth,
            const int pool_size, const int strides) {
     int out_len = input_len / strides;
     for (int start_index = 0; start_index < input_len; start_index += strides) {
@@ -60,13 +60,13 @@ void max1d(const int *data, int *map_out, const int input_len, const int input_d
                 if (mem2d(data, input_len, w_j, start_index + w_i) > maximum)
                     maximum = mem2d(data, input_len, w_j, start_index + w_i);
             }
-            mem2d(map_out, out_len, w_j, start_index / strides) = maximum;
+            mem2d(map_out, out_len, w_j, start_index / strides) = (short) maximum;
         }
     }
 }
 
-void batch_normalization(const int *data, const int *gamma, const int *beta, const int *mean, const int *var,
-                         int *map_out, const int input_len, const int input_depth){
+void batch_normalization(const short *data, const short *gamma, const short *beta, const short *mean, const short *var,
+                         short *map_out, const int input_len, const int input_depth){
     for (int start_index = 0; start_index < input_len; start_index ++) {
         for (int w_j = 0; w_j < input_depth; w_j++) {
             int normalized = MUL(mem2d(data, input_len, w_j, start_index) - mean[w_j] , var[w_j]);
@@ -75,13 +75,13 @@ void batch_normalization(const int *data, const int *gamma, const int *beta, con
     }
 }
 
-void relu(const int *data, int *map_out, int input_len) {
+void relu(const short *data, short *map_out, int input_len) {
     for (int i = 0; i < input_len; i++)
         map_out[i] = (data[i] < 0) ? 0 : data[i];
 }
 
 
-void save_file(int *data, char *filename, int input_len) {
+void save_file(short *data, char *filename, int input_len) {
     FILE *fp;
     fp = fopen(filename, "w");
     for (int wi = 0; wi < input_len; wi++)
@@ -90,16 +90,16 @@ void save_file(int *data, char *filename, int input_len) {
     fclose(fp);
 }
 
-void forward_propagation(int *data) {
+void forward_propagation(short *data) {
     int depth_size[6] = {2, 128, 128, 128, 100, 1};
     int map_size[6] = {1024, 256, 64, 16, 1, 1};
-    int *filter;
-    int *bias;
-    int *layer_in = (int *) data;
+    short *filter;
+    short *bias;
+    short *layer_in = (short *) data;
     save_file(data, "input_data.txt", 2048);
 
     for (int block = 0; block < 3; block++) {
-        int* conv1d_out = (int *) malloc(map_size[block] * depth_size[block + 1] * sizeof(int));
+        short* conv1d_out = (short *) malloc(map_size[block] * depth_size[block + 1] * sizeof(short));
         filter = conv1d_w[block];
         bias = conv1d_b[block];
         conv1d(layer_in, filter, conv1d_out, bias, 3, map_size[block], depth_size[block], map_size[block],
@@ -115,14 +115,14 @@ void forward_propagation(int *data) {
 
         relu(conv1d_out, conv1d_out, map_size[block] * depth_size[block + 1]);
 
-        int* max_out = (int *) malloc(map_size[block+1] * depth_size[block + 1] * sizeof(int));
+        short* max_out = (short *) malloc(map_size[block+1] * depth_size[block + 1] * sizeof(short));
         max1d(conv1d_out, max_out, map_size[block], depth_size[block+1], 4, 4);
         sprintf(out_filename, "max_pool_%d.txt", block);
         save_file(max_out, out_filename, map_size[block + 1] * depth_size[block + 1]);
         layer_in = max_out;
     }
     for (int fc_index=0; fc_index<2; fc_index++) {
-        int *fully_connected = malloc(map_size[4+fc_index] * depth_size[4+fc_index] * sizeof(int));
+        short *fully_connected = malloc(map_size[4+fc_index] * depth_size[4+fc_index] * sizeof(short));
         conv1d(layer_in, dense_w[fc_index], fully_connected, dense_b[fc_index], map_size[3+fc_index],
                map_size[3+fc_index], depth_size[3+fc_index],
                map_size[4+fc_index], depth_size[4+fc_index], map_size[3+fc_index],
